# expression: optimize BCE on the builtinFromUnixTime2ArgSig