Signed-off-by: Weizhen Wang <wangweizhen@pingcap.com>

<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: ref #34669

Problem Summary:

Go is a memory safe language. In array/slice element indexing and subslice operations, Go runtime will check whether or not the involved indexes are out of range. If an index is out of range, a panic will be produced to prevent the invalid index from doing harm. This is called bounds check. Bounds checks make our code run safely, on the other hand, they also make our code run a little slower.

### What is changed and how it works?

```
make vectorized-bench VB_FILE=Time VB_FUNC=builtinFromUnixTime2ArgSig 
```

Before

```go
goos: darwin
goarch: amd64
pkg: github.com/pingcap/tidb/expression
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkVectorizedBuiltinTimeFuncGenerated
BenchmarkVectorizedBuiltinTimeFuncGenerated-16    	1000000000	         0.3045 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinTimeFunc
BenchmarkVectorizedBuiltinTimeFunc-16             	1000000000	         0.06184 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/pingcap/tidb/expression	7.986s
```

After

```
goos: darwin
goarch: amd64
pkg: github.com/pingcap/tidb/expression
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkVectorizedBuiltinTimeFuncGenerated
BenchmarkVectorizedBuiltinTimeFuncGenerated-16          1000000000               0.2601 ns/op          0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinTimeFunc
BenchmarkVectorizedBuiltinTimeFunc-16                   1000000000               0.05196 ns/op         0 B/op          0 allocs/op
PASS
ok      github.com/pingcap/tidb/expression      5.989s
```

**17% performance improvement**

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [x] Integration test

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
None
```