@hawkingrei 2022-05-16
update benchmark report

```
benchmark                                          old ns/op     new ns/op     delta
BenchmarkVectorizedBuiltinTimeFuncGenerated-16     0.18          0.19          +3.19%
BenchmarkVectorizedBuiltinTimeFuncGenerated-16     0.19          0.19          +0.85%
BenchmarkVectorizedBuiltinTimeFuncGenerated-16     0.18          0.19          +6.57%
BenchmarkVectorizedBuiltinTimeFuncGenerated-16     0.19          0.19          +1.15%
BenchmarkVectorizedBuiltinTimeFuncGenerated-16     0.19          0.19          +3.52%
BenchmarkVectorizedBuiltinTimeFuncGenerated-16     0.20          0.19          -1.28%
BenchmarkVectorizedBuiltinTimeFuncGenerated-16     0.18          0.20          +7.12%
BenchmarkVectorizedBuiltinTimeFuncGenerated-16     0.19          0.19          -0.57%
BenchmarkVectorizedBuiltinTimeFuncGenerated-16     0.20          0.20          +1.53%
BenchmarkVectorizedBuiltinTimeFuncGenerated-16     0.20          0.20          -0.55%
BenchmarkVectorizedBuiltinTimeFunc-16              0.04          0.04          +1.83%
BenchmarkVectorizedBuiltinTimeFunc-16              0.04          0.04          +14.12%
BenchmarkVectorizedBuiltinTimeFunc-16              0.04          0.04          +16.38%
BenchmarkVectorizedBuiltinTimeFunc-16              0.04          0.04          -2.64%
BenchmarkVectorizedBuiltinTimeFunc-16              0.04          0.04          -0.41%
BenchmarkVectorizedBuiltinTimeFunc-16              0.04          0.04          -13.60%
BenchmarkVectorizedBuiltinTimeFunc-16              0.04          0.04          +0.05%
BenchmarkVectorizedBuiltinTimeFunc-16              0.04          0.04          +13.30%
BenchmarkVectorizedBuiltinTimeFunc-16              0.04          0.04          +3.00%
BenchmarkVectorizedBuiltinTimeFunc-16              0.04          0.04          +0.79%
```