@hawkingrei 2022-08-12
> It does improve the perf, but it is also ugly. I don't think we should merge this.
> 
> We should consider some alternatives, .e.g. moving some logic into column, instead in the expression package. Then BCE could be eliminated while the outer package does not have many assumptions on the internal implementation.
> 
> @hawkingrei What do you think about?

I think it is a test to show the possibility that we can optimize BCE to get some performance improvement. but it is really ugly. I think we can improve golang SSA to make the Golang compiler optimize BCE. but it is a hard task.