@xhebox 2022-08-12
> I think it is a test to show the possibility that we can optimize BCE to get some performance improvement. but it is really ugly.

You could put the benchmark and the one line core change into your issue. Thus this PR could be closed. I am cleaning some stale PRs and issues.

> I think we can improve golang SSA to make the Golang compiler optimize BCE. but it is a hard task.

I doubt if it is possible. Golang is not good at inlining, or they don't make much effort. It will be very difficult to optimize BCE across functions.