@xhebox 2022-08-15
> You could put the benchmark and the one line core change into your issue. Thus this PR could be closed. I am cleaning some stale PRs and issues.

Closing. Put the example into the tracking issue.